Move automated task helpers to Utils, reorganise question-specific rake tasks
